+++ lib/stigg/models/v1/coupon_list_params.rb
@@ -41,8 +41,8 @@ module Stigg
         # @!attribute status
         #   Filter by coupon status. Supports comma-separated values for multiple statuses
         #
-        #   @return [String, nil]
-        optional :status, String
+        #   @return [Array<Symbol, Stigg::Models::V1::CouponListParams::Status>, nil]
+        optional :status, -> { Stigg::Internal::Type::ArrayOf[enum: Stigg::V1::CouponListParams::Status] }
 
         # @!attribute type
         #   Filter by coupon type (FIXED or PERCENTAGE)
@@ -61,7 +61,7 @@ module Stigg
         #
         #   @param limit [Integer] Maximum number of items to return
         #
-        #   @param status [String] Filter by coupon status. Supports comma-separated values for multiple statuses
+        #   @param status [Array<Symbol, Stigg::Models::V1::CouponListParams::Status>] Filter by coupon status. Supports comma-separated values for multiple statuses
         #
         #   @param type [Symbol, Stigg::Models::V1::CouponListParams::Type] Filter by coupon type (FIXED or PERCENTAGE)
         #
@@ -102,6 +102,16 @@ module Stigg
           #   @param lt [Time] Less than the specified createdAt value
           #
           #   @param lte [Time] Less than or equal to the specified createdAt value
+        end
+
+        module Status
+          extend Stigg::Internal::Type::Enum
+
+          ACTIVE = :ACTIVE
+          ARCHIVED = :ARCHIVED
+
+          # @!method self.values
+          #   @return [Array<Symbol>]
         end
 
         # Filter by coupon type (FIXED or PERCENTAGE)