+++ lib/stigg/models/v1/feature_list_features_params.rb
@@ -35,8 +35,9 @@ module Stigg
         # @!attribute feature_type
         #   Filter by feature type. Supports comma-separated values for multiple types
         #
-        #   @return [String, nil]
-        optional :feature_type, String
+        #   @return [Array<Symbol, Stigg::Models::V1::FeatureListFeaturesParams::FeatureType>, nil]
+        optional :feature_type,
+                 -> { Stigg::Internal::Type::ArrayOf[enum: Stigg::V1::FeatureListFeaturesParams::FeatureType] }
 
         # @!attribute limit
         #   Maximum number of items to return
@@ -47,14 +48,16 @@ module Stigg
         # @!attribute meter_type
         #   Filter by meter type. Supports comma-separated values for multiple types
         #
-        #   @return [String, nil]
-        optional :meter_type, String
+        #   @return [Array<Symbol, Stigg::Models::V1::FeatureListFeaturesParams::MeterType>, nil]
+        optional :meter_type,
+                 -> { Stigg::Internal::Type::ArrayOf[enum: Stigg::V1::FeatureListFeaturesParams::MeterType] }
 
         # @!attribute status
         #   Filter by feature status. Supports comma-separated values for multiple statuses
         #
-        #   @return [String, nil]
-        optional :status, String
+        #   @return [Array<Symbol, Stigg::Models::V1::FeatureListFeaturesParams::Status>, nil]
+        optional :status,
+                 -> { Stigg::Internal::Type::ArrayOf[enum: Stigg::V1::FeatureListFeaturesParams::Status] }
 
         # @!method initialize(id: nil, after: nil, before: nil, created_at: nil, feature_type: nil, limit: nil, meter_type: nil, status: nil, request_options: {})
         #   @param id [String] Filter by entity ID
@@ -65,13 +68,13 @@ module Stigg
         #
         #   @param created_at [Stigg::Models::V1::FeatureListFeaturesParams::CreatedAt] Filter by creation date using range operators: gt, gte, lt, lte
         #
-        #   @param feature_type [String] Filter by feature type. Supports comma-separated values for multiple types
+        #   @param feature_type [Array<Symbol, Stigg::Models::V1::FeatureListFeaturesParams::FeatureType>] Filter by feature type. Supports comma-separated values for multiple types
         #
         #   @param limit [Integer] Maximum number of items to return
         #
-        #   @param meter_type [String] Filter by meter type. Supports comma-separated values for multiple types
+        #   @param meter_type [Array<Symbol, Stigg::Models::V1::FeatureListFeaturesParams::MeterType>] Filter by meter type. Supports comma-separated values for multiple types
         #
-        #   @param status [String] Filter by feature status. Supports comma-separated values for multiple statuses
+        #   @param status [Array<Symbol, Stigg::Models::V1::FeatureListFeaturesParams::Status>] Filter by feature status. Supports comma-separated values for multiple statuses
         #
         #   @param request_options [Stigg::RequestOptions, Hash{Symbol=>Object}]
 
@@ -110,6 +113,39 @@ module Stigg
           #   @param lt [Time] Less than the specified createdAt value
           #
           #   @param lte [Time] Less than or equal to the specified createdAt value
+        end
+
+        module FeatureType
+          extend Stigg::Internal::Type::Enum
+
+          BOOLEAN = :BOOLEAN
+          NUMBER = :NUMBER
+          ENUM = :ENUM
+
+          # @!method self.values
+          #   @return [Array<Symbol>]
+        end
+
+        module MeterType
+          extend Stigg::Internal::Type::Enum
+
+          NONE = :None
+          FLUCTUATING = :FLUCTUATING
+          INCREMENTAL = :INCREMENTAL
+
+          # @!method self.values
+          #   @return [Array<Symbol>]
+        end
+
+        module Status
+          extend Stigg::Internal::Type::Enum
+
+          NEW = :NEW
+          SUSPENDED = :SUSPENDED
+          ACTIVE = :ACTIVE
+
+          # @!method self.values
+          #   @return [Array<Symbol>]
         end
       end
     end